@@ -19,10 +19,10 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": ">=7.4.0"
+        "php": ">=8.1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.0",
+        "phpunit/phpunit": "^10.0",
         "php-coveralls/php-coveralls": "^2.0",
         "vimeo/psalm": "^5.4",
         "rector/rector": "^1.0.0"
