@@ -23,10 +23,9 @@ return static function (RectorConfig $rectorConfigurator): void {
     $rectorConfigurator->import(SetList::PHP_72);
     $rectorConfigurator->import(SetList::PHP_73);
     $rectorConfigurator->import(SetList::PHP_74);
-    //$containerConfigurator->import(SetList::PHP_80);
-    //$containerConfigurator->import(SetList::PHP_81);
+    $rectorConfigurator->import(SetList::PHP_80);
+    $rectorConfigurator->import(SetList::PHP_81);
     $rectorConfigurator->import(SetList::DEAD_CODE);
-    //$rectorConfigurator->import(SetList::PSR_4);
     $rectorConfigurator->import(SetList::TYPE_DECLARATION);
     
     $skipables = [
@@ -35,11 +34,6 @@ return static function (RectorConfig $rectorConfigurator): void {
         \Rector\CodingStyle\Rector\Encapsed\EncapsedStringsToSprintfRector::class,
         \Rector\DeadCode\Rector\PropertyProperty\RemoveNullPropertyInitializationRector::class,
     ];
-    
-    if(PHP_MAJOR_VERSION < 8) {
-        
-        $skipables[] = \Rector\DeadCode\Rector\ClassMethod\RemoveUnusedPromotedPropertyRector::class;
-    }
     
     $rectorConfigurator->skip($skipables);
 };
